Upgrade emotion to v10, remove helper, dynamic functions
Emotion 10 is beta so using that now. Some changes with that:

- You have to explicitly use a babel pragma to get the `css` prop.
- Server rendering just works! Nothing special to do there except render like normal.
- Using the `css` function is now just declarative. It doesn't actually create a CSS class. So creating static blobs of style is cool.

Because that last point and because I wasn't actually doing anything dynamic here, I just switched to a simple object pointing to static blobs of style. For anything dynamic, I'll probably mix in dynamic styles in the render function.