Prevent codeMirror.operation throwing
This fixes an issue where an error is thrown because we try to access `operation` of off `this.codeMirror` when `this.codeMirror === null`.

It's set to `null` in `removeCodeMirrorEditor`, which is called when the component unmounts or `updateEditor` is called (when a particular condition is met), and presumably that happens synchronously, and then the `operation` method attempts to be called in the next frame, due to `requestAnimationFrame`.