Pass typeName/parentTypeName through props
Previously, I was using adding `fieldTypeName` to the field and using that to determine the type name for the `renderTag` method. That didn't work for some components, like `array-item`, which are rendering child fields for their parents. In that case, the fieldTypeName would be something like `String`, but that was incorrect in the context of `renderTag`. Now, I'm manually passing around `typeName` and for those "middle" cases, I'm passing through `parentTypeName`, so it's clear that the `field` is for a child, but the `parentTypeName` is for the parent.