All new styling for demo fields
I killed off all the bootstrap styles. And I removed all dependencies on the old utility classes. Instead, styling can be done completely through a plugin using the `renderTag` method. In general I cleaned up a lot of the styles so things look a little better.

I did make a breaking change to the ordering of array controls. It makes more sense to have those after the value to match the way objects work, and this makes styling with flexbox more intuitive. So this forces a major bump. Also fixed some smaller quirks here and there. And I removed the transition wrapper for the choice dropdown for now, since it didn't work anyway.