Get docs building again; clean up dependencies, scripts
Because the docs pages depend on an API, we have to start up an API server, build the docs, and then kill the server.

When adding `gh-pages` for publishing docs, I noticed rollup was missing (I had removed it before), and adding that, I found some problems there. Got that mostly working but will probably do some follow-up commits to do a proper esm build.