renderTag has to pass through children
This takes away the magic conversion of children to `React.Fragment`. Instead, children are passed through as arguments and have to be handled by `renderTag`. The reason for this is there's too much potential for breaking changes where a plugin expected a component to render multiple children as an array and instead just gets a single element. This does put some extra burden on the consumer if they want to make `renderTag` performant, but in most cases, `...` should be fine. And otherwise, it's considered a pretty low level hook anyway.