Split rollup babel config from next babel config
Rollup and Next.js got into a fight over the babel config. So I added a plugin to Next that lets me point to an alternate babel config. That way rollup can own the root babel config (which is appropriate since that belongs to Formatic's code), and Next can be on its own. Perhaps Next needs to move to a sub-directory, but I'll save that for later.