Make sure all CSS gets added to docs in the proper order
The previous hack for CSS was unfortunately putting the CSS out of order. This is still hacky, but it's a better hack. It uses the styles.chunk.css file, so it's not manually recreating what was imported, and it's subclassing Head so that it can guarantee the order the CSS is included on the page.