Fix debounce problem
We made some improvements to debouncing in 2.0.4, but unfortunately, we caused a problem when we took out the flag that determined if we were in debouncing mode. That flag was used to determine when it was appropriate to update the state, so without it, we would make it impossible to load in dynamic replace choices. Now putting back the flag to fix that.