Add IntegerField/IntegerInput
This is just a simple example of a field that coerces its output value. This should probably be coupled with validation at some point so that what you're typing isn't blocked, but rather it becomes a validation error. That's for later though.

The spec currently describes using a schema for coercion, but I'm unsure if that's necessary:

https://github.com/zapier/formatic/pull/124#discussion_r278377148

So this is just an example of what I was describing there.