Even more performance ReactiveValue with metadata
Now, ReactiveValue wrapper instances are shared across hooks with the same property key. This means nested reactive values should perform even better.

Also added a way to subscribe to changes to property types of values. This is more performant than Object.keys(value) since the value would change every time a proprty changes, but the property types only change when the type of a property changes or a property is added.