Fix bug trying ta tag non-existing CodeMirror instance.
Sometimes we see this error in the console:

https://cdn.zapier.com/storage/photos/69d375365045c5fd6e9c7e2a68c43905.png

This happens when the debounced method to tag the CodeMirror editor tries
tag this.codeMirror after the component has unmounted and this.codeMirror
has been set to null on componentWillUnmount.

To fix, bail out of the method if `this.codeMirror` is not defined.